MERGE: -c 36625 -> /trunk/Sapphire
Improve formatting of log messages

(lsass: r36626)

+++ lsass/server/api/provider.c
@@ -108,8 +108,8 @@ cleanup:
 error:
 
     LSA_LOG_ERROR_API_FAILED(hServer, dwError,
-        "run provider specific request (request code = %ld, provider = '%s')",
-        (long)dwIoControlCode, LSA_SAFE_LOG_STRING(pszProvider));
+        "run provider specific request (request code = %u, provider = '%s')",
+        dwIoControlCode, LSA_SAFE_LOG_STRING(pszProvider));
 
     *pdwOutputBufferSize = 0;
     *ppOutputBuffer = NULL;
